squash a warning (size_t != int)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2223 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/lua51/lstrlib.c
@@ -1056,7 +1056,7 @@ static int str_split (lua_State *L) {
   }
   p = malloc(l+1);
   if (p==NULL) {
-	fprintf(stderr, "fatal: memory exhausted (malloc of %u bytes).\n",l+1);
+	fprintf(stderr, "fatal: memory exhausted (malloc of %u bytes).\n",(int)(l+1));
 	exit(EXIT_FAILURE);
   }
   strcpy(p,s);
